Fix compiler warnings from glib→Objectively migration
- r_media.c, s_media.c: HashTableHashFunc return type uint32_t→size_t
- g_ai_node.c, g_ai_main.c: remove const from locals passed to addElement/insertElementAtIndex
- g_ai_node.c: %d→%zu for size_t format arg
- g_entity.c, sv_editor.c: VectorElement dereference fixes (T** assigned to T*)
- main.c: PRIu32→%zu for size_t console block count

+++ src/client/renderer/r_media.c
@@ -262,7 +262,7 @@ void R_EndLoading(void) {
 /**
  * @brief Computes the hash value for a media entry by name and type.
  */
-static uint32_t R_MediaHash(const void * key) {
+static size_t R_MediaHash(const void * key) {
   const r_media_t *media = key;
   uint32_t hash = 5381;
